requirements.txt
markdown-it-py==3.0.0
mdurl==0.1.2
Pygments==2.17.2
pyperclip==1.8.2
python-slugify==8.0.4
rich==13.7.1
text-unidecode==1.3
